Update species traits file
Removed verified_taxonomy data package in favor of just the .csv file describing species traits. This is b/c 1) the DP required a relative file directory, which wont work unless the entire DP is saved to the package, and 2) only the traits table is used in the package (for now). Also updated where this is used in the download function.